requirements.txt
matplotlib
networkx
scipy
flask
werkzeug
python-docx